--- app/src/main/java/com/example/androidtrainingproject/LogInRequest.kt
@@ -1,9 +1,0 @@
-package com.example.androidtrainingproject
-
-
-import com.google.gson.annotations.SerializedName
-
-data class LogInRequest(
-    @SerializedName("user")
-    val user: User
-)
