--- app/src/main/java/com/example/androidtrainingproject/LogInResponse.kt
@@ -1,9 +1,0 @@
-package com.example.androidtrainingproject
-
-
-import com.google.gson.annotations.SerializedName
-
-data class LogInResponse(
-    @SerializedName("user")
-    val user: UserX
-)
