+++ app/src/main/java/com/example/androidtrainingproject/MainActivity.kt
@@ -3,14 +3,16 @@ package com.example.androidtrainingproject
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
-import android.util.Patterns
-import android.widget.Button
-import android.widget.EditText
 import android.widget.Toast
 import com.example.androidtrainingproject.databinding.ActivityMainBinding
+import com.example.androidtrainingproject.model.request.LogInRequest
+import com.example.androidtrainingproject.model.response.LogInResponse
+import com.example.androidtrainingproject.model.LoginRequestParams
 import retrofit2.Call
 import retrofit2.Callback
 
+
+private const val TAG = "MainActivity_d"
 
 class MainActivity : AppCompatActivity() {
     lateinit var binding:ActivityMainBinding
@@ -27,11 +29,16 @@ class MainActivity : AppCompatActivity() {
 
     private fun getInfo() {
 
-        email=binding.gmail.text.toString()
-        password=binding.password.text.toString()
+
+
         binding.press.setOnClickListener()
         {
 
+            email = "var123@gmail.com"
+            password = "123123123"
+
+    /*        email=binding.gmail.text.toString()
+            password=binding.password.text.toString()
 
            Toast.makeText(this,"${email.length}",Toast.LENGTH_SHORT).show()
             Toast.makeText(this,"${password.length}",Toast.LENGTH_SHORT).show()
@@ -55,16 +62,14 @@ class MainActivity : AppCompatActivity() {
             {
                 Toast.makeText(this,"Enter 6 digit password",Toast.LENGTH_SHORT).show()
                 return@setOnClickListener
-            }
-                    Service.logininterface.getData(LogInRequest(User(email, password)))
+            }*/
+                    RetrofitClient.logininterface.getData(LogInRequest(LoginRequestParams(email, password)))
                         .enqueue(object : Callback<LogInResponse> {
                             override fun onResponse(
                                 call: Call<LogInResponse>,
                                 response: retrofit2.Response<LogInResponse>
                             ) {
-                                Log.d("success", "cool")
-                                Toast.makeText(this@MainActivity, "success", Toast.LENGTH_SHORT)
-                                    .show()
+                                Log.d(TAG, "LoginData: "+response.body().toString())
 
                             }
