+++ app/src/main/java/com/example/androidtrainingproject/model/response/LogInResponse.kt
@@ -1,0 +1,10 @@
+package com.example.androidtrainingproject.model.response
+
+
+import com.example.androidtrainingproject.model.User
+import com.google.gson.annotations.SerializedName
+
+data class LogInResponse(
+    @SerializedName("user")
+    val user: User
+)
