+++ readMe.md
@@ -1,0 +1,8 @@
+#TODO: 
+- 
+
+# LEARNT: 
+
+### RetofitClient.kt
+- lazy{}
+-
